Fixes stuff & tests other

+++ UIElement.cpp
@@ -20,8 +20,11 @@ int UIElement::getDrawOrder()
 
 void UIElement::setShapePixelSize(sf::Vector2i screensize)
 {
-	shape.setPosition(sf::Vector2f(screensize.x * (bounds.left / 100), screensize.y * (bounds.top / 100)));
-	shape.setSize(sf::Vector2f(screensize.x * (bounds.width / 100), screensize.y * (bounds.height / 100)));
+	shape.setPosition(sf::Vector2f(screensize.x * ((float)bounds.left / 100), screensize.y * ((float)bounds.top / 100)));
+	shape.setSize(sf::Vector2f(screensize.x * ((float)bounds.width / 100), screensize.y * ((float)bounds.height / 100)));
+
+	//std::cout << shape.getPosition().x << ", " << shape.getPosition().y << "\n";
+	//std::cout << shape.getSize().x << ", " << shape.getSize().y << "\n";
 }
 
 UIElement::~UIElement()
